+++ R/viewer.R
@@ -2,9 +2,12 @@
 #'
 #' View map tiles in the browser using leaflet.
 #'
-#' This function opens \code{preview.html} in a web browser. This file displays map tiles in a Leaflet widget.
-#' The file is created when \code{tile} is called to generate the map tiles, unless \code{viewer = FALSE}.
-#' Alternatively, it is created (or re-created) subsequent to tile creation using \code{tile_viewer}.
+#' This function opens \code{preview.html} in a web browser. This file displays
+#' map tiles in a Leaflet widget.
+#' The file is created when \code{tile} is called to generate the map tiles,
+#' unless \code{viewer = FALSE}.
+#' Alternatively, it is created (or re-created) subsequent to tile creation
+#' using \code{tile_viewer}.
 #'
 #' @param tiles character, directory where tiles are stored.
 #'
@@ -18,8 +21,9 @@
 view_tiles <- function(tiles){
   file <- file.path(dirname(tiles), "preview.html")
   if(!file.exists(file)){
-    stop(paste("Cannot find preview.html.", "Tiles may have been generated with `viewer = FALSE`.",
-                 "Use `tile_viewer` to create preview.html."))
+    stop(paste("Cannot find preview.html.",
+               "Tiles may have been generated with `viewer = FALSE`.",
+               "Use `tile_viewer` to create preview.html."))
   }
   utils::browseURL(file)
 }
@@ -28,34 +32,51 @@ view_tiles <- function(tiles){
 #'
 #' Create an HTML file that displays a tile preview using Leaflet.
 #'
-#' This function creates a file \code{preview.html} adjacent to the \code{tiles} base directory.
-#' When loaded in the browser, this file displays map tiles from the adjacent folder.
-#' For example, if tiles are stored in \code{project/tiles}, this function creates \code{project/preview.html}.
+#' This function creates a file \code{preview.html} adjacent to the
+#' \code{tiles} base directory.
+#' When loaded in the browser, this file displays map tiles from the adjacent
+#' folder.
+#' For example, if tiles are stored in \code{project/tiles}, this function
+#' creates \code{project/preview.html}.
 #'
-#' By default, \code{tile} creates this file. The only reasons to call \code{tile_viewer} directly after producing map tiles are:
+#' By default, \code{tile} creates this file. The only reasons to call
+#' \code{tile_viewer} directly after producing map tiles are:
 #' (1) if \code{viewer = FALSE} was set in the call to \code{tile},
-#' (2) if \code{tile} was called multiple times, e.g., for different batches of zoom levels, and thus the most recent call did not use the full zoom range,
+#' (2) if \code{tile} was called multiple times, e.g., for different batches of
+#' zoom levels, and thus the most recent call did not use the full zoom range,
 #' or (3) \code{preview.html} was deleted for some other reason.
 #'
-#' If calling this function directly, ensure that the min and max zoom, and original image pixel dimensions if applicable, match the generated tiles.
-#' These arguments are passed to \code{tile_viewer} automatically when called within \code{tile}, based on the source file provided to \code{tile}.
+#' If calling this function directly, ensure that the min and max zoom, and
+#' original image pixel dimensions if applicable, match the generated tiles.
+#' These arguments are passed to \code{tile_viewer} automatically when called
+#' within \code{tile}, based on the source file provided to \code{tile}.
 #'
 #' @param tiles character, directory where tiles are stored.
 #' @param zoom character, zoom levels full range. Example format: \code{"3-7"}.
-#' @param width \code{NULL} (default) for geospatial map tiles. The original image width in pixels for non-geographic, simple CRS tiles.
-#' @param height \code{NULL} (default) for geospatial map tiles. The original image height in pixels for non-geographic, simple CRS tiles.
-#' @param georef logical, for non-geographic tiles only. If \code{viewer = TRUE}, then the Leaflet widget in \code{preview.html} will add map markers with coordinate labels on mouse click to assist with georeferencing of non-geographic tiles.
-#' @param ... additional optional arguments include \code{format = "tms"} if necessary, and \code{lng} and \code{lat} for setting the view longitude and latitude. These three arguments only apply to geographic tiles. Viewer centering is \code{0, 0} by default.
+#' @param width \code{NULL} (default) for geospatial map tiles. The original
+#' image width in pixels for non-geographic, simple CRS tiles.
+#' @param height \code{NULL} (default) for geospatial map tiles. The original
+#' image height in pixels for non-geographic, simple CRS tiles.
+#' @param georef logical, for non-geographic tiles only.
+#' If \code{viewer = TRUE}, then the Leaflet widget in \code{preview.html} will
+#' add map markers with coordinate labels on mouse click to assist with
+#' georeferencing of non-geographic tiles.
+#' @param ... additional optional arguments include \code{format = "tms"} if
+#' necessary, and \code{lng} and \code{lat} for setting the view longitude and
+#' latitude. These three arguments only apply to geographic tiles. Viewer
+#' centering is \code{0, 0} by default.
 #'
 #' @return nothing is returned, but a file is written to disk.
 #' @export
 #' @seealso \code{\link{view_tiles}}, \code{\link{tile}}
 #'
 #' @examples
-#' tile_viewer(file.path(tempdir(), "tiles"), "3-7") # requires an existing tile set
-tile_viewer <- function(tiles, zoom, width = NULL, height = NULL, georef = TRUE, ...){
+#' tile_viewer(file.path(tempdir(), "tiles"), "3-7") # requires existing tiles
+tile_viewer <- function(tiles, zoom, width = NULL, height = NULL,
+                        georef = TRUE, ...){
   index_js <- .index_js(tiles, zoom, width, height, georef, ...)
-  if(inherits(index_js, "character")) index_js <- paste(.raster_coords, index_js, sep = "\n\n")
+  if(inherits(index_js, "character"))
+    index_js <- paste(.raster_coords, index_js, sep = "\n\n")
   html <- .preview_html(index_js)
   writeLines(html, file.path(dirname(tiles), "preview.html"))
   invisible()
@@ -132,13 +153,15 @@ tile_viewer <- function(tiles, zoom, width = NULL, height = NULL, georef = TRUE,
 .index_js <- function(tiles, zoom, width, height, georef, ...){
   zoom <- rep(strsplit(as.character(zoom), "-")[[1]], length = 2)
   if(is.null(width) | is.null(height)){
-    fmt <- if(!is.null(list(...)$format) && list(...)$format == "tms") "tms" else "xyz"
+    fmt <- if(!is.null(list(...)$format) && list(...)$format == "tms")
+      "tms" else "xyz"
     lng <- if(is.null(list(...)$lng)) 0 else list(...)$lng
     lat <- if(is.null(list(...)$lat)) 0 else list(...)$lat
     x <- paste0("function init () {
       var mymap = L.map('map').setView([", lat, ", ", lng, "], ", zoom[1], ");
-      L.tileLayer('", basename(tiles), "/{z}/{x}/{y}.png', { minZoom: ", zoom[1], ", maxZoom: ",
-      zoom[2], ifelse(fmt == "tms", ", tms: true", ""), " }).addTo(mymap)
+      L.tileLayer('", basename(tiles), "/{z}/{x}/{y}.png', { minZoom: ",
+                zoom[1], ", maxZoom: ", zoom[2],
+                ifelse(fmt == "tms", ", tms: true", ""), " }).addTo(mymap)
     }")
   } else {
     markers <- ifelse(georef, "var layerBounds = L.layerGroup([
@@ -163,7 +186,8 @@ tile_viewer <- function(tiles, zoom, width = NULL, height = NULL, georef = TRUE,
       var rc = new L.RasterCoords(map, img)
       map.setMaxZoom(", zoom[2], ")
       map.setView(rc.unproject([img[0] / 2, img[1] / 2]), 2)\n", markers,
-      "L.tileLayer('", basename(tiles), "/{z}/{x}/{y}.png', { noWrap: true }).addTo(map)
+      "L.tileLayer('", basename(tiles),
+      "/{z}/{x}/{y}.png', { noWrap: true }).addTo(map)
     }")
   }
   x