+++ tests/testthat/test-viewer.R
@@ -11,6 +11,8 @@ test_that("viewer creation runs", {
   x <- file.path(tempdir(), "tiles")
   expect_is(tile_viewer(x, "0-3"), "NULL")
   expect_is(tile_viewer(x, "0-3", lng = 20, lat = 30, format = "tms"), "NULL")
-  expect_is(tile_viewer(x, "0-3", width = 1000, height = 1000, georef = FALSE), "NULL")
-  expect_is(tile_viewer(x, "0-3", width = 1000, height = 1000, georef = TRUE), "NULL")
+  expect_is(tile_viewer(x, "0-3", width = 1000, height = 1000, georef = FALSE),
+            "NULL")
+  expect_is(tile_viewer(x, "0-3", width = 1000, height = 1000, georef = TRUE),
+            "NULL")
 })